Disable UNK Alerts by default
This PR adds an option to FORCE ENABLE the UNK alerts that arise due to Amazon Cloudwatch throttling. We have seen a lot of alerts that are generated and clear due to throttling limits, this should be a mode that is not enabled by default as it causes unnecessary noise. Ideally this should be enabled however practically it causes false alerts 99% of the time.